"test: move handler tests into top-level tests package"
