+++ tests/Healthz_test.go
@@ -1,16 +1,18 @@
-package handlers
+package tests
 
 import (
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	h "devops-valgfag/handlers"
 )
 
 func TestHealthz_OK(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, "/healthz", nil)
 	rec := httptest.NewRecorder()
 
-	Healthz(rec, req)
+	h.Healthz(rec, req)
 
 	if rec.Code != http.StatusOK {
 		t.Fatalf("expected status 200, got %d", rec.Code)
