+++ tests/search_page_test.go
@@ -1,10 +1,12 @@
-package handlers
+package tests
 
 import (
 	"database/sql"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	h "devops-valgfag/handlers"
 )
 
 func closeDB(t *testing.T, db *sql.DB) {
@@ -21,7 +23,7 @@ func TestSearchPageHandler_NoQuery_OK(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, "/", nil)
 	rec := httptest.NewRecorder()
 
-	SearchPageHandler(rec, req)
+	h.SearchPageHandler(rec, req)
 
 	if rec.Code != http.StatusOK {
 		t.Fatalf("expected status 200, got %d", rec.Code)
@@ -38,7 +40,7 @@ func TestSearchPageHandler_WithQuery_OK(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, "/?q=hello&language=en", nil)
 	rec := httptest.NewRecorder()
 
-	SearchPageHandler(rec, req)
+	h.SearchPageHandler(rec, req)
 
 	if rec.Code != http.StatusOK {
 		t.Fatalf("expected status 200, got %d", rec.Code)
